Big changes.
No more JavaScript until I find an area where it is convenient to implement it in.  Thus, JQuery still stays.  Yes, this also means I've removed the entire JQuery OpenID-selector, because it's a lot more problematic than it's worth.  Just enter your damn OpenID URL or don't login at all.

Markdown now has the Pygments syntax highlighter implemented, shamelessly copied and pasted from the default example in the pygments code repo, with a few necessary modifications.  The formatting is: [sourcecode:lexer]code[/sourcecode], where lexer is the Pygments lexer, and code is the desired code to highlight.

Post adding now actually works!  No more hacking about manually with the Pylons Command Line to add a blog post, even if it wasn't such a bad idea anyway. This includes creation of categories on the spot, which I figure this doesn't deserve its own full-blown page and form, since categories are just titles and slugs (which are automatically generated.)  No authority for you.

Back when SUIT was really bad at looping, I had to convert every SQLAlchemy instance in a list to a dict before being able to loop through such values, and I would use "template helper" functions such as make_categories, make_posts, etc. - Since this is not the case anymore, they have been completely abolished.

This also means a new helper function, convert_datetime is now available, which runs .strftime() on datetime objects with the configuration specified formatting string.  If a string is passed to the function (when used from a SUIT template,) dateutil will attempt to convert it to a datetime object automatically to format it.  In addition, thanks to Brandon, alternating table rows/zebra tables can be made directly in SUIT with no need for non-flexible boilerplate code.  A [replace] rule was added, which may be removed depending on its usefulness.

It can be used as such:
    [replace what="best" with="worst"]I am the best.[/replace]

FormEncode's OpenId validator sucks a lot, because it doesn't verify if the provided value is an actual OpenID end-point, nor does it allow the not_empty argument, which makes handling the field a huge hassle.  I went ahead and brewed up a simple validator that run's the discover function from the openid.yadis.discover module on the provided value and return an eror.  Not the best idea ever, but a lot more convenient.

That is all.